#1730 remove NavTopBar component
